move UIHelper into util
